fix: reorganizar tipos
